requirements.txt
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.18.0
lxml>=5.3.0
aiohttp>=3.10.0
pandas>=2.2.0
tqdm>=4.66.0
python-dotenv>=1.0.0
